Hide FFI cbindgen behind a feature flag (#333)
